feat(zms): Added support for headless mode
These changes enable the execution of ZMS code without a REQUEST object, which is passed by Zope from a client in normal mode. This is needed to be able to access the objects stored in ZODB directly using Python code - without a running Zope server.

At the University of Bern, this is being used to implement a REST API that utilizes FastAPI as a developer-centred framework instead of Zope to access the ZMS content objects in ZODB. At the moment, only those ZMS methods that are required for the use cases of UniBE Mobile App have been changed - others will follow gradually.

More information: https://idasm-unibe-ch.github.io/zms-fastapi/